Integration Testing 1: Product Search filter
The product index has a search filter input box named product['product_sku']  (autogenerated based on the filterModelName + attribute ) enabling the searching for the product  sku or stock keeping unit.

The search uses [](https://www.w3schools.com/howto/howto_js_filter_table.asp)

How the coding works:
The product/index DataColumn filterInputAtributes includes the id that will appear in the input box productt['product_sku] that will be identified by the product.js function   $(document).on('click', '#product_filters_submit', function ().

The input boxes value is passed to the ProductController/search function using  the json data from the product.js file passed in the url query.  The product/index GridView urlQueryParameters array includes the attribute name. The search function captures this by means of the queryParams php function which gathers the json parameters into an array.

if the sku is found via the repository it is returned with a success flag of 1 which is used by the product..js function to determine if the tableFunction should be run eliminating those rows that DO NO MATCH the search criteria.