Connect Client Entity to FormModel construct
In order for the data from the client entity not to be lost when updated, the formModel has to be hydrated, or filled with data from the client entity. This is done by passing the client entity data to the formModel constructor in the form of an array since the data can either be passed from the currently entering form to the controller using the limited queryParams function which is in the form of an array  or the recommended $request->getParsedBody() ?? [];   function which is also in the form of an array, each component of the array maintaining their type and therefore satisfying static analysis.