Testing 7: PageSizeLimiter Widget filters on inv/index and product/index
3 non-javascript filters Invoice Number, Client, and Total introduced on the index.
The relevant currency is show in the header rather than next to the content column value.
A default page size limit picker widget is displayed and included in the summary template now so that it is not necessary to go into the settings view to select a suitable page size given the number of items in the list.
The productController has been updated as well.