Bugfix InvController inv/guest
Multiple Route Arguments have to be at the end of the function's argument structure and not in the middle of a group of arguments.

Query param from paginator is a string e.g. ?page=2
RouteArgument from status buttons is an int e.g. pressing 'all' button ...http://invoice.myhost/invoice/client_invoices/page/1/status/0

Therefore use mixed for 'page' i.e. $pageMixed and type cast when needed to either a string or an int.