Installation Testing 8: Move more variables into LayoutViewInjection
Include stopLoggingIn and stopSigningUp variable into the LayoutViewInjection
Include these variables into main.php
Replace currentRoute with RouteArguments in InvoiceController.php
The $debugMode variable now is linked to the .env file. Alter the debugMode in the .env file now instead of the LayoutViewInjection.
If the $_ENV['BUILD_DATABASE']  environment variable is linked directly in the config/common params.php file it results in performance degradation so leave the setitng on 'false' for the moment.

Next: Adjusting the logo height and width parameters in the companyPrivate Entity.
The logo has been incorporated under the Invoice_web and the Quote_web files. Now to incorporate a qrCode image.