Installation Testing 12: Introduce Portugese Brazilian language
Update the steps to include a new language under the Views...Setting...Google Translate.
If the user sets a language before logging in, the package will redirect to invoice with the language chosen and not to the default language english and therefore not requiring an additional click on the dropdown language icon. Refer to AuthController.php