Filter Supler Globals
Super Global values are filtered with the php filter_input command.

An optional autoload_wo_filter_input.php has been created.

This change is a personal preference of mine and mainly to avoid the annoying Netbeans message - Do not access Super Globals directly -  and the fact that we are using a few, though arguably not many, filter_input functions  and also to get a bit more experience using the filter_input command.

With the package running, test by going to the .env file and change the YII_DEBUG value to false and the debug_menu will disappear.