Select All and Multiple Copy Invoice Checkbox Button
Select all javascipt checkbox working for inv/index...used with following buttons:
1. Multiple recurring button - mark invoices as recurring with stop and run button. Use with a cron if necessary.... still to be built in. Exploring possible use of crontab as mentioned in the blog-api demo.
2. Multiple sent button - mark invoices as sent so they appear on the client login
3. Multiple copy invoice button - sort the inv/index according to a client group and copy them using a specific date or if the  date is left blank it will use the current day. Useful for monthly batch invoicing.
If peppol is used, include the invoice item allowance or charge in the  invoice item (product or task) subtotal amount. A separate line will appear on the invoice for either the allowance or charge.

Disable the '+'  or 'burger button' on the invoice line item relevant to individual allowances and charges if peppol has not been enabled in settings.

Included the ability to delete all invoices for testing purposes with a function
inv/flush that deletes relevant tables which have foreign keys first in order to avoid foreign key constraint violations further down in other relaated tables that  need to be deleted.  The function appears under settings.

The line item subTotal now includes individual peppol allowances or charges .... if peppol is chosen to be used in order to generate an electronic invoice.

Removed yiisoft/yii-view from composer since yiisoft/yii-view-renderer replaces it.