Login Testing
When a user signs up, an email verification link is sent to the user to click on. If the  user does not click on this link they should not be able to login. The user can now only login now if the status of the user via the userinv table is active i.e. a value of 1. The status becomes active if they click on the link within the allotted 3600 seconds within their email account.

See the extension table userinv (of user) where the status field is included.
Also see the UserInvController extension Controller.