General Testing 3: Bugfix: No Invoice being created.
No invoice being created because user is not active most probably because the admin created them and forgot to make them active even though clients were assigned to them. Ideal solution: Have the user signup by internet and this will not occur.

New message.

Even if a user has several clients associated with it, no invoices will be creatable if this user whom they are associated with is not active.  Manually have to go to Invoice User Account under setting and make this user active.

Note: If the user signs up automatically by email their userinv account is automatically made active and they are given a client.