General Testing 17: Default list limit Button
The pageSizeLimiter button in  the summary template of each index can now be clicked on by the user and takes the user to the setting 'default_list_limit' so that the default can be adjusted without having to look for this button under settings.
Bugfix: Clicking on pdf in index exposed json result. Use pre-existing  dashboard void function instead directing user to a new browser tab in a blank page.
All indexes have been adjusted to allow the editing of the default list limit.