Github Oauth Functional
25th December 2024

Github Oauth2 functional. A Github user is authenticated, signed up, and assigned a client automatically without disclosing passwords.

A new oAuthService function oauthLogin has been created which does not require a password since authentication is already provided by Github.

In summary, the user is found by means of their login i.e. username with code

    Auth\\AuthService

    public function oauthLogin(string $login): bool
    {
        $user = $this->userRepository->findByLoginWithAuthIdentity($login);

        if ($user === null) {
            return false;
        }

        return $this->currentUser->login($user->getIdentity());
    }

and using the user identity, the current user is logged in.

Step 1. When the user chooses to 'Continue to Github', Github returns an access token using

$oAuthTokenType = $this->github->fetchAccessToken($request, $code, $params = \[\])

, and we use this token in a header request, to get information from the user by means of the Github Client in fork rossaddison/yii-auth-client

$this->github->getCurrentUserJsonArray($oAuthTokenType);

The auth/callbackGithub function concatenates a user using:

 $login = 'github'.(string)$githubId.$githubLogin;

a. 'github' ... to distinguish it from other Identity Providers.

b. The Github Id which is an integer and naturally unique.

c. and the Github 'login' or username which is perhaps not that unique and may have been used more than once by the user when registering with other Identity Providers e.g LinkedIn

Step 2. A random string password is built but will never be used and is hashed.

Step 3. The authService attempts to 'first time login' with this 'login' and deactivate the hour token ...otherwise a new user, and userinv is created, directing the user to login with their Identity Provider.

Step 4. The user must click on the 'please click here within the next hour to make active' button, to avoid our 'Github Access Token' from expiring.

Like the email verification token, our Github Access Token is set at an hour or 3600 seconds, allowing the user to activate and login within the hour and is unrelated to the Application's Github Access Token received from Github in step 1.