Facebook Testing
Clicking on cancel button when connecting to facebook gives rise to MissingRequiredArgumentException from the error parameters i.e callbackFacebook?error=access_denied&error_code=200&error_description=Permissions+error&error_reason=user_denied&state=...which are not normally in the redirectUrl.  To avoid this issue, add null to queries code and state and include error, error_code, and error_reason in query parameters that are returned from facebook to the returnUrl and catch the specific error with an if statement instead of a catch statement.

if (($errorCode == 200) && ($error == 'access_denied') && ($errorReason == 'user_denied')) {
                return $this->redirectToUserCancelledOauth2();