Yandex OAuth2.0 introduced
Upgrade X i.e. Twitter AuthController to use code_challenge_method 'S256' instead of 'plain', the same as Yandex which has been incorporated as well.
Session has been introduced in the AuthController for the purpose of recording the intial code_verifier value which is matched with the code_verifier value submitted when fetching the access_token.
Next: Adjust the AuthController so that state is recorded in session as well and compared with state received in the callback.