AuthController: blockInvalidState
Include a function blockInvalidState in all the AuthController callback clients which checks for csrf attacks and null session Auth state. Copilot was asked a question regarding this however the yii-auth-client repository already stores the state during the yii-auth-client oauth2 buildAuthUrl function as 'authState'. This value is compared with the state value returned by the IdentityProvider and if compared and they are not the same then there is likely a csrf attack. This potential has been included as well as potential for a null value since the AuthClient getState function is mixed so can return either a string value or null.