Modal Layout Simplification
Modal Layouts adjusted to reflect no 'aria-hidden' usage. Refactored yi-bootstrap/Modal has 'aria-hidden' issues and is not being used since these have to be suppressed with an 'inert' attribute via javascript. i.e. modal_layout_yii_bs.php and the modal itself is not rendering. See
https://github.com/yiisoft/yii-bootstrap5/issues/242

Invoices that have been marked as sent so that they appear on the customer side can be reverted to draft provided the 'disable_read_only' setting available only in Yii_DEBUG mode is adjusted to 1 by means of setting..view.

This currently has to be done manually by going into the non tab-indexed settings and adjusting the disable_read_only setting to 1.

I will be including a button on the related flash-message to make this easier next.