Introduce BaseController
A useful bat file for static analysis, requirement checking, and analysing outdated packages introduced with the help of AI in 5 minutes.

A BaseController for InvoiceController, InvController, QuoteController, and SalesOrderController, and a few other controllers have been introduced.
AI will subsequently help online to convert the others.

Rule of thumb: 'protected property' in BaseController allows the Controller, and methods that use the protected property  to be extended by child controllers that extend the BaseController.
Improve the inv\index's buttons relative to the disable.read.only setting and the is.read.only setting.
Buttons can be disabled with  'disabled' => 'disabled',
                                        'aria-disabled' => 'true',
                                        'style' => 'pointer-events:none' combination.