Credit Notes
Created an additional Group called Credit Note in the InvoiceController

Credit Notes can be generated if the original invoice has been sent.

The Credit Note is simply a 'reversed sent Invoice'.

Bugfix: When a Credit Note is created from an Invoice, this Invoice's creditinvoiceparent_id is set to the new Credit Notes id (effectively an invoice id)

I have not created a separate Credit Note entity

Payment methods can be made active. Only active payment methods should appear in the dropdowns. Discourage the use of cheques or cash.

Set the default payment method to online as dealing with cheques or cash complicates matters.