Testing Wonderful api
Replaced int with float since amount being sent via Wonderful was being truncated. e.g. £3.60 truncated to £3.00
The return url is functional but ref was being truncated because the route did not have it so included ref in the route.
A more general message is preferred on the completion page for Wonderful which is compiled in the PaymentInformationController function wonderful_complete.
Next: Refund a Credit Note if Payment has been made by the customer via Wonderful
Next: Include an Oauth2.0 payment provider e.g. Tink in the Open Banking selection process. Tink uses OIDC so this will be an interesting integration.
Current Payment Providers: ...src\Invoice\Setting\Trait\OpenBankProviders.php