Introduce Overall Packaging, Handling, Shipping
Peppol arrays have been updated to the 24th November 2024 arrays.
The $vat field for InvAllowanceCharge and InvItemAllowanceCharge have been adjusted to $vatOrTax so that non-vat operating
companies can include a gst tax or a tax similar to vat when working with handling, packing, and shipping fees.
I have included a field called Packhandleship_tax and packhandleship_total in the InvAmount entity which will reflect
the overall value of invoice allowances and charges when using the categories under Peppol. These categories are useful
in a non-peppol context and particularly a non-vat context hence the more flexible approach and renaming of $vat to
$vat_or_tax. The InvAmount entity has been adjusted to accomodate these two new totals which really are a net
   invallowancecharge field ... so not strictly just packing, handling, and shipping. The InvAllowanceCharge index
   has been upgraded to allow indexing and sorting.

An itemized breakdown of these fees can be accessed through a link on the view and these overall charges relating to the invoice are itemized in a separate section on the pdf invoice.

Step 1: Peppol ... Add Parent Invoice Allowance and Charges
Step 2: Under Options ... Add Overall Invoice Allowance and Charges
Step 3: Use the + sign on individual line items under the invoice view to input allowances or charges.

Refresh the view screen to see updated totals so allow redirects to complete once Inv Allowance Charges are entered.
The NumberHelper works independently of screen refreshes and redirects.

Next: Include some Codeception Tests related to the above and incorporate the above into xml e-invoicing. Validate with ecosio.