Update composer
The rossaddison/yii-auth-client dev-master will be syncing in with yiisoft/yii-auth-client due to pull request on 69 on yiisoft/yii-auth-client. A branch has been created called dev-invoice on rossaddison/yii-auth-client which will be maintained independently of rossaddison/yii-auth-client dev-master branch. A few useful changes e.g. a MakeFile which works with WSL can serve as an additional feature.