Reducing Code
A FormFields widget has been created for code reduction. Each function is tailored to specific entity needs.
Next: Custom Fields: Move basic functionality to the BaseController so that it code can also be reduced and try and implement some sort of validation on the custom fields. Also remove the 'NULL' appearing in the custom field.